Finalizando projeto

+++ src/main/java/com/github/Hugoftf/Spring/JPA/service/AutorService.java
@@ -1,8 +1,10 @@
 package com.github.Hugoftf.Spring.JPA.service;
 
 import com.github.Hugoftf.Spring.JPA.controller.dto.AutorDTO;
+import com.github.Hugoftf.Spring.JPA.exceptions.OperacaoNaoPermitida;
 import com.github.Hugoftf.Spring.JPA.model.Autor;
 import com.github.Hugoftf.Spring.JPA.repository.AutorRepository;
+import com.github.Hugoftf.Spring.JPA.repository.LivroRepository;
 import com.github.Hugoftf.Spring.JPA.service.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -15,10 +17,13 @@ public class AutorService {
 
     private AutorRepository autorRepository;
     private AutorValidator autorValidator;
+    private LivroRepository livroRepository;
 
-    public AutorService(AutorRepository autorRepository, AutorValidator autorValidator){
+    public AutorService(AutorRepository autorRepository, AutorValidator autorValidator,
+                        LivroRepository livroRepository){
         this.autorRepository = autorRepository;
         this.autorValidator = autorValidator;
+        this.livroRepository =  livroRepository;
     }
 
     public Autor salvar(Autor autor){
@@ -39,8 +44,14 @@ public class AutorService {
     }
 
     public void deletar(Autor autor) {
-         autorRepository.delete(autor);
+        if (possuiLivro(autor)){
+            throw new OperacaoNaoPermitida("Não é permitido excluir um autor com livros!");
+        }
+
+        autorRepository.delete(autor);
     }
+
+
 
     public List<Autor> encontrarPorNomeOuNacionalidade(String nome, String nacionalidade){
         if (nome != null && nacionalidade != null){
@@ -55,5 +66,9 @@ public class AutorService {
             return autorRepository.findByNacionalidade(nacionalidade);
         }
         return autorRepository.findAll();
+    }
+
+    private boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByIdAutor(autor);
     }
 }
